Fixes for PR #536 feedback
