+++ src/app/audit_log/pages/detail.html
@@ -45,8 +45,34 @@
           {% for change in detail.changes %}
             <tr class="{% if change.old_value.is_empty() %}added{% elif change.new_value.is_empty() %}removed{% else %}changed{% endif %}">
               <td>{{ change.field }}</td>
-              <td>{{ change.old_value }}</td>
-              <td>{{ change.new_value }}</td>
+              <td>
+                {% if change.old_refs.is_empty() %}
+                  {{ change.old_value }}
+                {% else %}
+                  {% for entity_ref in change.old_refs %}
+                    <div class="entity-ref">
+                      <a href="/audit-log?search={{ entity_ref.id_full }}">
+                        <abbr title="{{ entity_ref.id_full }}">{{ entity_ref.id_abbreviated }}</abbr>
+                      </a>
+                      {% if !entity_ref.description.is_empty() %}<span>{{ entity_ref.description }}</span>{% endif %}
+                    </div>
+                  {% endfor %}
+                {% endif %}
+              </td>
+              <td>
+                {% if change.new_refs.is_empty() %}
+                  {{ change.new_value }}
+                {% else %}
+                  {% for entity_ref in change.new_refs %}
+                    <div class="entity-ref">
+                      <a href="/audit-log?search={{ entity_ref.id_full }}">
+                        <abbr title="{{ entity_ref.id_full }}">{{ entity_ref.id_abbreviated }}</abbr>
+                      </a>
+                      {% if !entity_ref.description.is_empty() %}<span>{{ entity_ref.description }}</span>{% endif %}
+                    </div>
+                  {% endfor %}
+                {% endif %}
+              </td>
             </tr>
           {% endfor %}
         </tbody>
