+++ src/app/audit_log/pages/list.html
@@ -15,9 +15,9 @@
                 aria-label="{{ "audit_log.filter.event_type"|trans }}">
           <option value="">{{ "audit_log.filter.all_events"|trans }}</option>
           {% for group in event_types_by_category %}
-            {% let category_key = format!("audit_log.filter.category.{}", group.0) %}
+            {% let category_key = format!("audit_log.filter.category.{}", group.key) %}
             <optgroup label="{{ category_key.as_str() |trans }}">
-              {% for event_key in group.1.iter() %}
+              {% for event_key in group.event_types.iter() %}
                 {% let event_label_key = format!("audit_log.event.{}", event_key) %}
                 <option value="{{ event_key }}"
                         {% if filter.event_type.as_deref() == Some(event_key.as_ref()) %}selected{% endif %}>
@@ -28,6 +28,9 @@
           {% endfor %}
         </select>
         <button type="submit" class="button">{{ "audit_log.filter.apply"|trans }}</button>
+        {% if filter.is_active() %}
+          <a href="/audit-log" class="button secondary">{{ "audit_log.filter.reset"|trans }}</a>
+        {% endif %}
       </fieldset>
     </form>
   </section>
