+++ src/app/audit_log/structs/mod.rs
@@ -1,5 +1,9 @@
-mod audit_log_diff;
+mod audit_log_detail;
 mod audit_log_entry;
+mod entity_refs;
+mod event_info;
+mod event_payload;
+mod json_flatten;
 
-pub use audit_log_diff::AuditLogDetail;
+pub use audit_log_detail::AuditLogDetail;
 pub use audit_log_entry::AuditLogEntry;
