+++ src/app/authorised_agents/mod.rs
@@ -8,5 +8,5 @@ mod pages;
 mod structs;
 
 pub use forms::AuthorisedAgentForm;
-pub use pages::router;
+pub use pages::{AuthorisedAgentUpdatePath, router};
 pub use structs::{AuthorisedAgent, AuthorisedAgentId};
