+++ src/app/candidate_lists/mod.rs
@@ -9,7 +9,7 @@ mod pages;
 mod structs;
 
 pub use forms::{CandidateListCreateForm, CandidateListForm};
-pub use pages::router;
+pub use pages::{ViewCandidateListPath, router};
 pub use structs::{
     CandidateList, CandidateListId, CandidateListSummary, FullCandidateList, ListSubmitterForm,
 };
