+++ src/app/list_submitters/mod.rs
@@ -5,5 +5,5 @@ mod pages;
 mod structs;
 
 pub use forms::ListSubmitterForm;
-pub use pages::router;
+pub use pages::{ListSubmitterUpdatePath, router};
 pub use structs::{ListSubmitter, ListSubmitterData, ListSubmitterId};
