+++ src/app/persons/mod.rs
@@ -9,5 +9,5 @@ pub use forms::{
     AddressForm, PersonalDataFieldsForm, PersonalDataForm, RepresentativeFieldsForm,
     RepresentativeForm,
 };
-pub use pages::router;
+pub use pages::{UpdatePersonPath, router};
 pub use structs::{Person, PersonId, PersonPagination, PersonSort, PersonalData, Representative};
