+++ src/app/political_groups/mod.rs
@@ -6,6 +6,6 @@ mod steps;
 mod structs;
 
 pub use forms::PoliticalGroupForm;
-pub use pages::router;
+pub use pages::{PoliticalGroupUpdatePath, router};
 pub use steps::PoliticalGroupSteps;
 pub use structs::PoliticalGroup;
