+++ src/app/substitute_list_submitters/mod.rs
@@ -1,4 +1,4 @@
 //! Substitute list submitter management and related routes.
 mod pages;
 
-pub use pages::router;
+pub use pages::{SubstituteSubmitterUpdatePath, router};
